requirements.txt
#requirements.txt
requests
SQLAlchemy>=1.4
psycopg2-binary
python-dotenv
tenacity
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
xgboost>=1.6.0
jupyter>=1.0.0
